Store computedValue seperately from displayValue
This commit updates the `cell` to only store `value` and `displayValue`.
The latter is a new field that stores the stringified representation of
`computedValue`. That field now does not exist on the cell directly
anymore: it is now stored in a WeakMap inside of the worker that is
keyed by the cell object.

This restores the functionality of being able to have cell's contain
functions.